fix STACK display
